+++ CHANGELOG.md
@@ -4,13 +4,16 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 
 ## Stable
 
-[`v6.10.1 (2025-03-30)`](https://github.com/faasjs/faasjs/compare/v6.10.0...v6.10.1)
+[`v6.11.0 (2025-03-30)`](https://github.com/faasjs/faasjs/compare/v6.10.0...v6.11.0)
+
+- `@faasjs/server`
+  - [Feature] Add `middleware` to `Server`.
 
 - `faasjs`
   - [Fix] Fix `tsconfig.react.json`.
 
-- `@faasjs/server`
-  - [Feature] Add `middleware` to `Server`.
+- `@faasjs/react`
+  - [Fix] Fix reload error.
 
 [`v6.10.0 (2025-03-30)`](https://github.com/faasjs/faasjs/compare/v6.9.0...v6.10.0)